# mime: speed up ParseMediaType